Move GetTempDirectoryName to PathUtils for reuse in NuGet tests
