+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Cache/DataRegionMetricsTest.cs
@@ -49,7 +49,7 @@ namespace Apache.Ignite.Core.Tests.Cache
         private static readonly TimeSpan CheckpointFrequency = TimeSpan.FromSeconds(5);
 
         /** Temp dir for PDS. */
-        private static readonly string TempDir = TestUtils.GetTempDirectoryName();
+        private static readonly string TempDir = PathUtils.GetTempDirectoryName();
 
         /// <summary>
         /// Tests the memory metrics.
