+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Deployment/PeerAssemblyLoadingTest.cs
@@ -177,7 +177,7 @@ namespace Apache.Ignite.Core.Tests.Deployment
         {
             // Copy Apache.Ignite.exe and Apache.Ignite.Core.dll 
             // to a separate folder so that it does not locate our assembly automatically.
-            var folder = TestUtils.GetTempDirectoryName();
+            var folder = PathUtils.GetTempDirectoryName();
             foreach (var asm in new[] {typeof(IgniteRunner).Assembly, typeof(Ignition).Assembly})
             {
                 Assert.IsNotNull(asm.Location);
